Update SDK to use Session when initializing service clients
This change updates the way the SDK service clients are created.
Instead of taking an optional aws.Config pointer value the SDK now
requires a Session to be provided to the service clients when creating
instances. This session takes over for the global
`defaults.DefaultConfig`. Instead of being a global value the defaults
values are collected and set to a Session when `session.New()` is called.

To migrate to this change you'll need to add a session value to each
place a service cleint is created. Naively this can be just
`s3.New(session.New())` in place of `s3.New(nil)`. If aws.Config value
was passed in your code would become, `s3.New(session.New(), cfg)`. If
you used `default.DefaultConfig` you can create a new session, set the
config you used on the default config like `session.New(myCfg)` or
`sess.New(); sess.Config.LogLevel = aws.LogLevel(aws.Debug)`.

Examples:
```go
    // Create a session with the default config and request handlers.
    sess := session.New()

    // Create a session with a custom region
    sess := session.New(&aws.Config{Region: aws.String("us-east-1")})

    // Create a session, and add additional handlers for all service
    // clients created with the session to inherit.
    sess := session.New()
    sess.Handlers.Build.pushBack(func(r *request.Handlers) {
        // Log every request made and its payload
        logger.Println("Request: %s/%s, Payload: %s", r.ClientInfo.ServiceName, r.Operation, r.Params)
    })

    // Create a S3 client instance from a session
    sess := session.New()
    svc := s3.New(sess)

    // Create a S3 client with additional configuration
    svc := s3.New(sess, aws.NewConfig().WithRegion("us-west-2"))
```

In addition to this change there were a few other changes in this update.

* aws/service:Service was renamed to aws/client:Client.
* aws/service/serviceinfo:ServiceInfo was renamed to aws/client/metadata:ClientInfo
* aws:DefaultRetries was renamed to aws:UseServiceDefaultRetries
* aws/ec2metadata now takes a session and aws.Config like all other service clients
* aws/credentails/ec2rolecreds:EC2RoleProvider now requires an ec2metadata.EC2Metadata and will not default to creating one.
* service/s3/s3manager:Downloader now requires a S3 service client instance, and also takes functional options for setting the Downloader's options.
* service/s3/s3manager:Uploader now requires a S3 service client instance, and also takes functional options for setting the Uploader's options.